+++ packages/uring_api/src/_uring_api/uring_api_probe.c
@@ -248,6 +248,7 @@ static const UringApi_CAPI uring_api_capi_table = {
     UringApiCapi_CompletionClearUserData,
     UringApiCapi_RingWaitIdle,
     UringApiCapi_RingConstructSendAll,
+    UringApiCapi_CompletionTakeUserData,
 };
 
 int uring_api_export_capi(PyObject *module) {